Simplified and abstracted query execution
