+++ ghost/app.go
@@ -16,6 +16,7 @@ package ghost
 
 import (
 	"database/sql"
+	"strings"
 
 	"github.com/pressly/chi"
 	"github.com/spf13/afero"
@@ -35,7 +36,8 @@ type application struct {
 	Config config
 	//FileSystem is the main FileSystem
 	FileSystem afero.Fs
-	//Store
+	//Store is the data abstraction layer
+	//Normally your applications would interact with Store rather than DB or Cache
 	Store store
 }
 
@@ -52,25 +54,4 @@ func (a *application) Setup(configFileName string) {
 	//Initialise the filesysem
 	a.FileSystem = afero.NewOsFs()
 
-}
-
-type store struct
-
-func (s store) execute(queryString string) (string, error){
-
-		var dbResponse string
-	if err := App.DB.QueryRow(sqlString).Scan(&dbResponse); err != nil {
-		//Only one row is returned as JSON is returned by Postgres
-		//Empty result
-		if strings.Contains(err.Error(), "sql") {
-			return "", nil
-		}
-
-		//Else its a database error
-		return "", err
-
-	}
-
-	return dbResponse, nil
-	
 }
